requirements.txt
tensorflow==1.14
keras==2.2.5
numpy==1.16.5
scikit-image
matplotlib==3.1.1
seaborn==0.9.0
opencv-python==4.2.0.34
imageai==2.1.5
Pillow==6.2.0
pytesseract==0.3.4
ipython==7.13.0
scikit-learn
deskew
flask==1.1.1
flask_uploads
flask_dropzone